pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "n8n-harness"
version = "0.1.0"
description = "Thin attach-style harness for n8n: REST + MCP + a self-extending helpers.py."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
# Install with `uv tool install -e .[dev]` (or `pip install -e .[dev]`)
# to pick up pytest + anything else the test suite needs.
dev = [
    "pytest>=7",
]

[project.scripts]
n8n-harness = "run:main"

[tool.setuptools]
py-modules = ["run", "helpers", "admin"]

[tool.pytest.ini_options]
markers = [
    "gated_real: tests that hit the real n8n instance; skip if N8N_API_KEY unset",
]

